+++ test/asset_test.dart
@@ -19,7 +19,7 @@ void main() {
 
     test('constructor set the identifier correctly', () {
       const String id = 'SOME_ID';
-      Asset asset = Asset(id);
+      Asset asset = Asset(id, 100, 100);
       expect(
         asset.identifier,
         equals(id),
@@ -27,7 +27,7 @@ void main() {
     });
 
     test('thumbData can not have negative dimensions', () async {
-      Asset asset = Asset('_identifier');
+      Asset asset = Asset('_identifier', 50, 50);
 
       expect(
         asset.requestThumbnail(-100, 10),
@@ -37,6 +37,45 @@ void main() {
       expect(
         asset.requestThumbnail(10, -100),
         throwsArgumentError,
+      );
+    });
+
+    test('asset has correct dimensions', () async {
+      Asset asset = Asset('_identifier', 50, 40);
+
+      expect(
+        asset.originalWidth,
+        equals(50),
+      );
+
+      expect(
+        asset.originalHeight,
+        equals(40),
+      );
+    });
+
+    test('asset isLandscape and isPortrait work', () async {
+      Asset asset1 = Asset('_identifier', 50, 40);
+      Asset asset2 = Asset('_identifier', 40, 50);
+
+      expect(
+        asset1.isLandscape,
+        equals(true),
+      );
+
+      expect(
+        asset1.isPortrait,
+        equals(false),
+      );
+
+      expect(
+        asset2.isLandscape,
+        equals(false),
+      );
+
+      expect(
+        asset2.isPortrait,
+        equals(true),
       );
     });
   });